refactor(web): extract model service into pi-backend
