+++ web/packages/pi-backend/models.ts
@@ -1,0 +1,450 @@
+import type { ThinkingLevel } from "@earendil-works/pi-agent-core";
+import {
+  createAgentSessionServices,
+  getAgentDir,
+  resolveModelScopeWithDiagnostics,
+  type ModelRuntime,
+  type ScopedModel,
+  type SettingsManager,
+} from "@earendil-works/pi-coding-agent";
+import { getSupportedThinkingLevels, type Api, type Model } from "@earendil-works/pi-ai";
+import { projectTrustReloadOptions } from "../../lib/project-trust.ts";
+
+// ============================================================================
+// Model catalog cache (moved from lib/models-cache.ts verbatim).
+// ============================================================================
+export interface ModelsData {
+  models: Record<string, string>;
+  modelList: { id: string; name: string; provider: string }[];
+  defaultModel: { provider: string; modelId: string } | null;
+  thinkingLevels: Record<string, string[]>;
+  thinkingLevelMaps: Record<string, Record<string, string | null>>;
+  /** `provider/modelId` → thinking level pinned by an `enabledModels` `:level` suffix. */
+  thinkingLevelPins: Record<string, string>;
+  modelError?: string;
+  /** Warnings from resolving the `enabledModels` scope (e.g. a pattern matched nothing). */
+  modelScopeWarnings?: string[];
+}
+
+interface ModelsCacheState {
+  entries: Map<string, { data: ModelsData; expiresAt: number }>;
+  inFlight: Map<string, Promise<ModelsData>>;
+  generation: number;
+}
+
+declare global {
+  var __piModelsCacheState: ModelsCacheState | undefined;
+}
+
+const MODELS_CACHE_TTL_MS = 60_000;
+const MAX_MODELS_CACHE_ENTRIES = 32;
+// Never interpolate the caught error here; SDK errors can contain paths and provider details.
+const SAFE_MODEL_LOAD_FAILURE_MESSAGE = "Model list is temporarily unavailable. Check your configuration and try again.";
+
+function getModelsCacheState(): ModelsCacheState {
+  if (!globalThis.__piModelsCacheState) {
+    globalThis.__piModelsCacheState = {
+      entries: new Map(),
+      inFlight: new Map(),
+      generation: 0,
+    };
+  }
+  return globalThis.__piModelsCacheState;
+}
+
+export function invalidateModelsCache(): void {
+  const state = getModelsCacheState();
+  state.generation += 1;
+  state.entries.clear();
+  state.inFlight.clear();
+}
+
+export function withModelRuntimeError(data: ModelsData, modelError: string | undefined): ModelsData {
+  return modelError ? { ...data, modelError } : data;
+}
+
+export function withSafeModelLoadFailure(data: ModelsData): ModelsData {
+  return { ...data, modelError: SAFE_MODEL_LOAD_FAILURE_MESSAGE };
+}
+
+export function loadModelsWithCache(cwd: string, loader: () => Promise<ModelsData>): Promise<ModelsData> {
+  const state = getModelsCacheState();
+  const cached = state.entries.get(cwd);
+  if (cached) {
+    if (cached.expiresAt > Date.now()) return Promise.resolve(cached.data);
+    state.entries.delete(cwd);
+  }
+
+  const existingLoad = state.inFlight.get(cwd);
+  if (existingLoad) return existingLoad;
+
+  const generation = state.generation;
+  const loadPromise: Promise<ModelsData> = Promise.resolve()
+    .then(loader)
+    .then((data) => {
+      if (state.generation === generation && state.inFlight.get(cwd) === loadPromise) {
+        const now = Date.now();
+        for (const [key, entry] of state.entries) {
+          if (entry.expiresAt <= now) state.entries.delete(key);
+        }
+        while (state.entries.size >= MAX_MODELS_CACHE_ENTRIES) {
+          const oldestKey = state.entries.keys().next().value;
+          if (oldestKey === undefined) break;
+          state.entries.delete(oldestKey);
+        }
+        state.entries.set(cwd, { data, expiresAt: now + MODELS_CACHE_TTL_MS });
+      }
+      return data;
+    })
+    .finally(() => {
+      if (state.inFlight.get(cwd) === loadPromise) state.inFlight.delete(cwd);
+    });
+
+  state.inFlight.set(cwd, loadPromise);
+  return loadPromise;
+}
+
+// ============================================================================
+// Model scope (moved from lib/model-scope.ts verbatim).
+// ============================================================================
+const THINKING_LEVEL_SUFFIXES = new Set<ThinkingLevel>([
+  "off",
+  "minimal",
+  "low",
+  "medium",
+  "high",
+  "xhigh",
+  "max",
+]);
+
+/**
+ * Model scoping shared by the UI selector and AgentSession startup.
+ *
+ * The `enabledModels` setting uses the same syntax as pi's `--models` flag:
+ * globs matched with minimatch against `provider/modelId` or a bare `modelId`,
+ * fuzzy matching for non-glob patterns, plus an optional `:thinkingLevel` suffix
+ * (`anthropic/*:high`). Exact string comparison silently drops every model
+ * behind a pattern like `my-gateway/*` (#307), so delegate to pi's own resolver
+ * instead of reimplementing the matching rules here.
+ */
+
+export interface ModelScopeResult {
+  /** Models the UI should offer, in resolver order (all available when unscoped). */
+  visible: readonly Model<Api>[];
+  /** SDK-native scope retained for AgentSession model cycling and extensions. */
+  scopedModels: readonly ScopedModel[];
+  /** `provider/modelId` → thinking level pinned with a `:level` pattern suffix. */
+  thinkingLevelPins: Record<string, string>;
+  /** Resolver diagnostics, e.g. a pattern that matched no model. */
+  warnings: string[];
+}
+
+export interface InitialModelScopeOptions {
+  requestedModel?: { provider: string; modelId: string };
+  defaultModel?: { provider: string; modelId: string };
+  thinkingLevel?: ThinkingLevel;
+}
+
+export interface InitialModelScopeResult {
+  model?: Model<Api>;
+  thinkingLevel?: ThinkingLevel;
+  scopedModels: ScopedModel[];
+}
+
+function matchesModel(
+  model: { provider: string; id: string },
+  ref: { provider: string; modelId: string },
+): boolean {
+  return model.provider === ref.provider && model.id === ref.modelId;
+}
+
+function hasGlob(pattern: string): boolean {
+  return pattern.includes("*") || pattern.includes("?") || pattern.includes("[");
+}
+
+function exactReferenceMatches(pattern: string, models: readonly Model<Api>[]): Model<Api>[] {
+  const normalized = pattern.toLowerCase();
+  const canonical = models.filter(
+    (model) => `${model.provider}/${model.id}`.toLowerCase() === normalized,
+  );
+  if (canonical.length > 0) return canonical;
+  return models.filter((model) => model.id.toLowerCase() === normalized);
+}
+
+function assertNoAmbiguousExactPatterns(
+  patterns: readonly string[],
+  models: readonly Model<Api>[],
+): void {
+  for (const pattern of patterns) {
+    if (hasGlob(pattern)) continue;
+
+    let matches = exactReferenceMatches(pattern, models);
+    if (matches.length === 0) {
+      const colonIndex = pattern.lastIndexOf(":");
+      const suffix = colonIndex >= 0 ? pattern.slice(colonIndex + 1) : "";
+      if (THINKING_LEVEL_SUFFIXES.has(suffix as ThinkingLevel)) {
+        matches = exactReferenceMatches(pattern.slice(0, colonIndex), models);
+      }
+    }
+
+    if (matches.length > 1) {
+      const references = matches
+        .map((model) => `${model.provider}/${model.id}`)
+        .sort()
+        .join(", ");
+      throw new Error(
+        `Ambiguous enabledModels entry "${pattern}" matches multiple models: ${references}. Use provider/modelId.`,
+      );
+    }
+  }
+}
+
+/**
+ * Resolve the visible model list for `patterns`.
+ *
+ * Falls back to every available model when no patterns are configured or when
+ * the patterns resolve to nothing, so a stale or typo'd setting can never leave
+ * the UI without any selectable model.
+ */
+export async function resolveVisibleModels(
+  modelRuntime: ModelRuntime,
+  patterns: string[] | undefined,
+): Promise<ModelScopeResult> {
+  const cleaned = (patterns ?? []).map((pattern) => pattern.trim()).filter(Boolean);
+  if (cleaned.length === 0) {
+    return {
+      visible: await modelRuntime.getAvailable(),
+      scopedModels: [],
+      thinkingLevelPins: {},
+      warnings: [],
+    };
+  }
+
+  const available = await modelRuntime.getAvailable();
+  assertNoAmbiguousExactPatterns(cleaned, available);
+  const snapshotRuntime = {
+    getAvailable: async () => available,
+  } as ModelRuntime;
+  const { scopedModels, diagnostics } = await resolveModelScopeWithDiagnostics(cleaned, snapshotRuntime);
+  const warnings = diagnostics.map((diagnostic) => diagnostic.message);
+  if (scopedModels.length === 0) {
+    return {
+      visible: available,
+      scopedModels: [],
+      thinkingLevelPins: {},
+      warnings,
+    };
+  }
+
+  // `anthropic/*:high` pins a thinking level on every model the glob matched.
+  // pi applies the pin of the model a new session starts with; report them all
+  // so the client can look up whichever model it pre-selects.
+  const thinkingLevelPins: Record<string, string> = {};
+  for (const scoped of scopedModels) {
+    if (scoped.thinkingLevel) {
+      thinkingLevelPins[`${scoped.model.provider}/${scoped.model.id}`] = scoped.thinkingLevel;
+    }
+  }
+  return {
+    visible: scopedModels.map((scoped) => scoped.model),
+    scopedModels,
+    thinkingLevelPins,
+    warnings,
+  };
+}
+
+/**
+ * Select the model and thinking level used to create a new AgentSession.
+ *
+ * This mirrors pi's startup rule: prefer an explicit selection, otherwise use
+ * the saved default when it is in scope, then the first resolver-ordered model.
+ * A scoped-model thinking pin is applied unless the caller supplied an explicit
+ * thinking level.
+ */
+export function selectInitialModelScope(
+  scope: ModelScopeResult,
+  options: InitialModelScopeOptions = {},
+): InitialModelScopeResult {
+  const requestedRef = options.requestedModel;
+  const defaultRef = options.defaultModel;
+  const requested = requestedRef
+    ? scope.visible.find((model) => matchesModel(model, requestedRef))
+    : undefined;
+  if (requestedRef && !requested) {
+    throw new Error(
+      `Model is not available in the enabled scope: ${requestedRef.provider}/${requestedRef.modelId}`,
+    );
+  }
+
+  const requestedScoped = requested
+    ? scope.scopedModels.find((scoped) => scoped.model === requested
+      || matchesModel(scoped.model, { provider: requested.provider, modelId: requested.id }))
+    : undefined;
+  const defaultScoped = !requested && defaultRef
+    ? scope.scopedModels.find((scoped) => matchesModel(scoped.model, defaultRef))
+    : undefined;
+  const fallbackScoped = !requested ? (defaultScoped ?? scope.scopedModels[0]) : undefined;
+  const defaultVisible = !requested && !fallbackScoped && defaultRef
+    ? scope.visible.find((model) => matchesModel(model, defaultRef))
+    : undefined;
+  const selectedModel = requested ?? fallbackScoped?.model ?? defaultVisible;
+  const scopedSelection = requestedScoped ?? fallbackScoped;
+  const thinkingLevel = options.thinkingLevel ?? scopedSelection?.thinkingLevel;
+
+  return {
+    ...(selectedModel ? { model: selectedModel } : {}),
+    ...(thinkingLevel ? { thinkingLevel } : {}),
+    scopedModels: [...scope.scopedModels],
+  };
+}
+
+// ============================================================================
+// Startup preferences (moved from lib/startup-preferences.ts verbatim).
+// ============================================================================
+export interface ExplicitStartupPreferences {
+  model?: { provider: string; modelId: string };
+  thinkingLevel?: ThinkingLevel;
+}
+
+export interface EffectiveStartupPreferences {
+  model?: { provider: string; modelId: string };
+  thinkingLevel: ThinkingLevel;
+  supportsThinking: boolean;
+}
+
+/**
+ * Persist explicit browser selections without re-running AgentSession setters.
+ *
+ * The session constructor already records the effective model and thinking
+ * level. Calling setModel()/setThinkingLevel() again would append duplicate
+ * session entries and emit duplicate extension events.
+ */
+export async function persistExplicitStartupPreferences(
+  settingsManager: SettingsManager,
+  explicit: ExplicitStartupPreferences,
+  effective: EffectiveStartupPreferences,
+): Promise<{ modelDefaultChanged: boolean }> {
+  if (!explicit.model && !explicit.thinkingLevel) {
+    return { modelDefaultChanged: false };
+  }
+
+  let modelDefaultChanged = false;
+
+  if (
+    explicit.model
+    && effective.model
+    && explicit.model.provider === effective.model.provider
+    && explicit.model.modelId === effective.model.modelId
+  ) {
+    settingsManager.setDefaultModelAndProvider(
+      effective.model.provider,
+      effective.model.modelId,
+    );
+    modelDefaultChanged = true;
+  }
+
+  if (
+    explicit.thinkingLevel
+    && (effective.supportsThinking || effective.thinkingLevel !== "off")
+  ) {
+    settingsManager.setDefaultThinkingLevel(effective.thinkingLevel);
+  }
+
+  await settingsManager.flush();
+  return { modelDefaultChanged };
+}
+
+// ============================================================================
+// Model catalog service (moved from app/api/models/route.ts loadModels, minus
+// cwd transport validation which stays in the route) + the getModels entry.
+// ============================================================================
+const modelNameCollator = new Intl.Collator(undefined, { numeric: true, sensitivity: "base" });
+
+function compareModelEntries(
+  a: { id: string; name: string; provider: string },
+  b: { id: string; name: string; provider: string },
+): number {
+  return modelNameCollator.compare(a.name || a.id, b.name || b.id)
+    || modelNameCollator.compare(a.provider, b.provider)
+    || modelNameCollator.compare(a.id, b.id);
+}
+
+async function loadModels(cwd: string): Promise<ModelsData> {
+  const nameMap = new Map<string, string>();
+  let modelList: { id: string; name: string; provider: string }[] = [];
+  let defaultModel: { provider: string; modelId: string } | null = null;
+  const thinkingLevels: Record<string, string[]> = {};
+  const thinkingLevelMaps: Record<string, Record<string, string | null>> = {};
+
+  const agentDir = getAgentDir();
+  // Gate untrusted project extensions: enumerating models still imports and
+  // runs a repository's .pi/extensions factories, so honor project trust here
+  // too (see lib/project-trust.ts, #236).
+  const trustReloadOptions = projectTrustReloadOptions(cwd, agentDir);
+  const services = await createAgentSessionServices({
+    cwd,
+    agentDir,
+    ...(trustReloadOptions ? { resourceLoaderReloadOptions: trustReloadOptions } : {}),
+  });
+  const modelError = services.modelRuntime.getError();
+  const settings: SettingsManager = services.settingsManager;
+  // `enabledModels` supports globs and fuzzy patterns, so resolve it the same
+  // way the CLI does instead of comparing pattern strings literally (#307).
+  const scope = await resolveVisibleModels(
+    services.modelRuntime,
+    settings.getEnabledModels(),
+  );
+  const { visible, thinkingLevelPins, warnings } = scope;
+  modelList = visible.map((m) => ({
+    id: m.id,
+    name: m.name,
+    provider: m.provider,
+  })).sort(compareModelEntries);
+  for (const m of visible) {
+    const key = `${m.provider}:${m.id}`;
+    nameMap.set(key, m.name);
+    thinkingLevels[key] = getSupportedThinkingLevels(m);
+    if (m.thinkingLevelMap) thinkingLevelMaps[key] = m.thinkingLevelMap;
+  }
+
+  const defaultProvider = settings.getDefaultProvider();
+  const defaultModelId = settings.getDefaultModel();
+  const initial = selectInitialModelScope(scope, {
+    ...(defaultProvider && defaultModelId
+      ? { defaultModel: { provider: defaultProvider, modelId: defaultModelId } }
+      : {}),
+  });
+  if (initial.model) {
+    defaultModel = { provider: initial.model.provider, modelId: initial.model.id };
+  }
+
+  return withModelRuntimeError(
+    {
+      models: Object.fromEntries(nameMap),
+      modelList,
+      defaultModel,
+      thinkingLevels,
+      thinkingLevelMaps,
+      thinkingLevelPins,
+      ...(warnings.length > 0 ? { modelScopeWarnings: warnings } : {}),
+    },
+    modelError,
+  );
+}
+
+const EMPTY_MODELS: ModelsData = {
+  models: {},
+  modelList: [],
+  defaultModel: null,
+  thinkingLevels: {},
+  thinkingLevelMaps: {},
+  thinkingLevelPins: {},
+};
+
+export async function getModels(cwd: string): Promise<ModelsData> {
+  try {
+    return await loadModelsWithCache(cwd, () => loadModels(cwd));
+  } catch {
+    return withSafeModelLoadFailure(EMPTY_MODELS);
+  }
+}
